Restructure and abstract
